Login usuario

+++ src/main/java/net/daw/service/UsuarioService.java
@@ -38,22 +38,31 @@ public class UsuarioService {
         ob = oRequest.getParameter("ob");
     }
 
+    protected Boolean checkPermission(String ob, String strMethodName) {
+        UsuarioBean oUsuarioBean = (UsuarioBean) oRequest.getSession().getAttribute("user");
+        return oUsuarioBean != null;
+    }
+
     public ReplyBean get() throws Exception {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            UsuarioBean oUsuarioBean = oUsuarioDao.get(id, 1);		
-            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
-            oReplyBean = new ReplyBean(200, oGson.toJson(oUsuarioBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: get method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission(ob, "get")) {
+            try {
+                Integer id = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                UsuarioBean oUsuarioBean = oUsuarioDao.get(id, 1);
+                Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+                oReplyBean = new ReplyBean(200, oGson.toJson(oUsuarioBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: get method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
 
         return oReplyBean;
@@ -64,17 +73,21 @@ public class UsuarioService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer id = Integer.parseInt(oRequest.getParameter("id"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            int iRes = oUsuarioDao.remove(id);
-            oReplyBean = new ReplyBean(200, Integer.toString(iRes));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: remove method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission(ob, "remove")) {
+            try {
+                Integer id = Integer.parseInt(oRequest.getParameter("id"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                int iRes = oUsuarioDao.remove(id);
+                oReplyBean = new ReplyBean(200, Integer.toString(iRes));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: remove method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
 
@@ -84,17 +97,21 @@ public class UsuarioService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            int registros = oUsuarioDao.getcount();
-            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
-            oReplyBean = new ReplyBean(200, oGson.toJson(registros));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission(ob, "getcount")) {
+            try {
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                int registros = oUsuarioDao.getcount();
+                Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+                oReplyBean = new ReplyBean(200, oGson.toJson(registros));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: getcount method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
 
         return oReplyBean;
@@ -105,20 +122,24 @@ public class UsuarioService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            String strJsonFromClient = oRequest.getParameter("json");
-            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
-            UsuarioBean oUsuarioBean = new UsuarioBean();
-            oUsuarioBean = oGson.fromJson(strJsonFromClient, UsuarioBean.class);
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            oUsuarioBean = oUsuarioDao.create(oUsuarioBean);
-            oReplyBean = new ReplyBean(200, oGson.toJson(oUsuarioBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: create method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission(ob, "create")) {
+            try {
+                String strJsonFromClient = oRequest.getParameter("json");
+                Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+                UsuarioBean oUsuarioBean = new UsuarioBean();
+                oUsuarioBean = oGson.fromJson(strJsonFromClient, UsuarioBean.class);
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                oUsuarioBean = oUsuarioDao.create(oUsuarioBean);
+                oReplyBean = new ReplyBean(200, oGson.toJson(oUsuarioBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: create method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
     }
@@ -128,20 +149,24 @@ public class UsuarioService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            String strJsonFromClient = oRequest.getParameter("json");
-            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
-            UsuarioBean oUsuarioBean = new UsuarioBean();
-            oUsuarioBean = oGson.fromJson(strJsonFromClient, UsuarioBean.class);
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            iRes = oUsuarioDao.update(oUsuarioBean);
-            oReplyBean = new ReplyBean(200, Integer.toString(iRes));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: update method: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission(ob, "update")) {
+            try {
+                String strJsonFromClient = oRequest.getParameter("json");
+                Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+                UsuarioBean oUsuarioBean = new UsuarioBean();
+                oUsuarioBean = oGson.fromJson(strJsonFromClient, UsuarioBean.class);
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                iRes = oUsuarioDao.update(oUsuarioBean);
+                oReplyBean = new ReplyBean(200, Integer.toString(iRes));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: update method: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
         return oReplyBean;
     }
@@ -150,20 +175,24 @@ public class UsuarioService {
         ReplyBean oReplyBean;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
-            Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
-            HashMap<String, String> hmOrder = ParameterCook.getOrderParams(oRequest.getParameter("order"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            ArrayList<UsuarioBean> alUsuarioBean = oUsuarioDao.getpage(iRpp, iPage, hmOrder, 1);
-            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
-            oReplyBean = new ReplyBean(200, oGson.toJson(alUsuarioBean));
-        } catch (Exception ex) {
-            throw new Exception("ERROR: Service level: get page: " + ob + " object", ex);
-        } finally {
-            oConnectionPool.disposeConnection();
+        if (this.checkPermission(ob, "getpage")) {
+            try {
+                Integer iRpp = Integer.parseInt(oRequest.getParameter("rpp"));
+                Integer iPage = Integer.parseInt(oRequest.getParameter("page"));
+                HashMap<String, String> hmOrder = ParameterCook.getOrderParams(oRequest.getParameter("order"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                ArrayList<UsuarioBean> alUsuarioBean = oUsuarioDao.getpage(iRpp, iPage, hmOrder, 1);
+                Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+                oReplyBean = new ReplyBean(200, oGson.toJson(alUsuarioBean));
+            } catch (Exception ex) {
+                throw new Exception("ERROR: Service level: get page: " + ob + " object", ex);
+            } finally {
+                oConnectionPool.disposeConnection();
+            }
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
         }
 
         return oReplyBean;
@@ -174,24 +203,66 @@ public class UsuarioService {
         ReplyBean oReplyBean = null;
         ConnectionInterface oConnectionPool = null;
         Connection oConnection;
-        try {
-            Integer cantidadUsuarios = Integer.parseInt(oRequest.getParameter("cant"));
-            oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
-            oConnection = oConnectionPool.newConnection();
-            Gson oGson = new Gson();
-            Generadorusuarios oGeneradorusuarios = new Generadorusuarios();
-            UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
-            UsuarioBean oUsuarioBean;
-            for (int i = 0; i < cantidadUsuarios; i++) {
-                oUsuarioBean = oGeneradorusuarios.generar();
-                oUsuarioDao.create(oUsuarioBean);
+        if (this.checkPermission(ob, "filldatabase")) {
+            try {
+                Integer cantidadUsuarios = Integer.parseInt(oRequest.getParameter("cant"));
+                oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+                oConnection = oConnectionPool.newConnection();
+                Gson oGson = new Gson();
+                Generadorusuarios oGeneradorusuarios = new Generadorusuarios();
+                UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+                UsuarioBean oUsuarioBean;
+                for (int i = 0; i < cantidadUsuarios; i++) {
+                    oUsuarioBean = oGeneradorusuarios.generar();
+                    oUsuarioDao.create(oUsuarioBean);
+                }
+                oReplyBean = new ReplyBean(200, oGson.toJson("{\"usuarios creados\":" + cantidadUsuarios + "}"));
+            } catch (Exception ex) {
+                oReplyBean = new ReplyBean(500,
+                        "ERROR: " + EncodingHelper.escapeQuotes(EncodingHelper.escapeLine(ex.getMessage())));
+            } finally {
+                oConnectionPool.disposeConnection();
             }
-            oReplyBean = new ReplyBean(200, oGson.toJson("{\"usuarios creados\":" + cantidadUsuarios + "}"));
-        } catch (Exception ex) {
-            oReplyBean = new ReplyBean(500,
-                    "ERROR: " + EncodingHelper.escapeQuotes(EncodingHelper.escapeLine(ex.getMessage())));
-        } finally {
-            oConnectionPool.disposeConnection();
+        } else {
+            oReplyBean = new ReplyBean(401, "Unauthorized");
+        }
+        return oReplyBean;
+    }
+
+    public ReplyBean login() throws Exception {
+        ReplyBean oReplyBean;
+        ConnectionInterface oConnectionPool = null;
+        Connection oConnection;
+        String strLogin = oRequest.getParameter("user");
+        String strPassword = oRequest.getParameter("pass");
+        oConnectionPool = ConnectionFactory.getConnection(ConnectionConstants.connectionPool);
+        oConnection = oConnectionPool.newConnection();
+        UsuarioDao oUsuarioDao = new UsuarioDao(oConnection, ob);
+        UsuarioBean oUsuarioBean = oUsuarioDao.login(strLogin, strPassword);
+        if (oUsuarioBean.getId() > 0) {
+            oRequest.getSession().setAttribute("user", oUsuarioBean);
+            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+            oReplyBean = new ReplyBean(200, oGson.toJson(oUsuarioBean));
+        } else {
+            oReplyBean = new ReplyBean(401, "Bad Authentication");
+        }
+        return oReplyBean;
+    }
+
+    public ReplyBean logout() throws Exception {
+        oRequest.getSession().invalidate();
+        return new ReplyBean(200, "OK");
+    }
+
+    public ReplyBean check() throws Exception {
+        ReplyBean oReplyBean;
+        UsuarioBean oUsuarioBean;
+        oUsuarioBean = (UsuarioBean) oRequest.getSession().getAttribute("user");
+        if (oUsuarioBean != null) {
+            Gson oGson = (new GsonBuilder()).excludeFieldsWithoutExposeAnnotation().create();
+            oReplyBean = new ReplyBean(200, oGson.toJson(oUsuarioBean));
+        } else {
+            oReplyBean = new ReplyBean(401, "No active session");
         }
         return oReplyBean;
     }
